test(core): split shared test helpers
