+++ packages/core/test/core/helpers/paths.ts
@@ -1,0 +1,3 @@
+import { resolve } from 'node:path';
+
+export const ROOT_NODE_MODULES = resolve(import.meta.dir, '../../../../..', 'node_modules');
